+++ .github/workflows/main.yml
@@ -43,6 +43,7 @@ jobs:
           node-version: ${{ matrix.node-version }}
       - run: sudo sqlite3 /Library/Application\ Support/com.apple.TCC/TCC.db "select * from access;" | grep kTCCServiceScreenCapture
       - run: sudo sqlite3 /Library/Application\ Support/com.apple.TCC/TCC.db "select * from access;"
+      - run: node temp.js
       # - run: npm install
       # - name: Download build
       #   uses: actions/download-artifact@v3
